+++ tests/compile_fail/dynify_with_unsupported_item.rs
@@ -1,5 +1,8 @@
 #[dynify::dynify]
-fn test() {}
+fn test1() {}
+
+#[dynify::dynify]
+fn test2() -> FakeImpl {}
 
 #[dynify::dynify]
 opaque_trait!();